remove Ignore from SingleRowJoinTest, and update based on comments
