+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/util/FlinkRelOptUtil.scala
@@ -19,11 +19,10 @@ package org.apache.flink.table.plan.util
 
 import org.apache.flink.api.common.operators.Order
 import org.apache.flink.table.api.TableException
-import org.apache.flink.table.functions.sql.internal.SqlAuxiliaryGroupAggFunction
 
 import org.apache.calcite.plan.RelOptUtil
 import org.apache.calcite.rel.RelFieldCollation.Direction
-import org.apache.calcite.rel.core.{Aggregate, AggregateCall, JoinInfo}
+import org.apache.calcite.rel.core.{AggregateCall, JoinInfo}
 import org.apache.calcite.rel.{RelFieldCollation, RelNode}
 import org.apache.calcite.rex.{RexLiteral, RexNode}
 import org.apache.calcite.sql.{SqlExplainLevel, SqlKind}
@@ -33,7 +32,6 @@ import org.apache.calcite.util.mapping.IntPair
 import java.io.{PrintWriter, StringWriter}
 import java.util
 
-import scala.collection.JavaConversions._
 import scala.collection.mutable
 
 object FlinkRelOptUtil {
@@ -110,59 +108,6 @@ object FlinkRelOptUtil {
     */
   def containsAccurateDistinctCall(aggCalls: Seq[AggregateCall]): Boolean = {
     aggCalls.exists(call => call.isDistinct && !call.isApproximate)
-  }
-
-  /**
-    * Check whether AUXILIARY_GROUP aggCalls is in the front of the given agg's aggCallList,
-    * and whether aggCallList contain AUXILIARY_GROUP when the given agg's groupSet is empty
-    * or the indicator is true.
-    * Returns AUXILIARY_GROUP aggCalls' args and other aggCalls.
-    *
-    * @param agg aggregate
-    * @return returns AUXILIARY_GROUP aggCalls' args and other aggCalls
-    */
-  def checkAndSplitAggCalls(agg: Aggregate): (Array[Int], Seq[AggregateCall]) = {
-    var nonAuxGroupCallsStartIdx = -1
-
-    val aggCalls = agg.getAggCallList
-    aggCalls.zipWithIndex.foreach {
-      case (call, idx) =>
-        if (call.getAggregation == SqlAuxiliaryGroupAggFunction) {
-          require(call.getArgList.size == 1)
-        }
-        if (nonAuxGroupCallsStartIdx >= 0) {
-          // the left aggCalls should not be AUXILIARY_GROUP
-          require(call.getAggregation != SqlAuxiliaryGroupAggFunction,
-            "AUXILIARY_GROUP should be in the front of aggCall list")
-        }
-        if (nonAuxGroupCallsStartIdx < 0 &&
-          call.getAggregation != SqlAuxiliaryGroupAggFunction) {
-          nonAuxGroupCallsStartIdx = idx
-        }
-    }
-
-    if (nonAuxGroupCallsStartIdx < 0) {
-      nonAuxGroupCallsStartIdx = aggCalls.length
-    }
-
-    val (auxGroupCalls, otherAggCalls) = aggCalls.splitAt(nonAuxGroupCallsStartIdx)
-    if (agg.getGroupCount == 0) {
-      require(auxGroupCalls.isEmpty,
-        "AUXILIARY_GROUP aggCalls should be empty when groupSet is empty")
-    }
-    if (agg.indicator) {
-      require(auxGroupCalls.isEmpty,
-        "AUXILIARY_GROUP aggCalls should be empty when indicator is true")
-    }
-
-    val auxGrouping = auxGroupCalls.map(_.getArgList.head.toInt).toArray
-    require(auxGrouping.length + otherAggCalls.length == aggCalls.length)
-    (auxGrouping, otherAggCalls)
-  }
-
-  def checkAndGetFullGroupSet(agg: Aggregate): Array[Int] = {
-    val (auxGroupSet, _) = checkAndSplitAggCalls(agg)
-    agg.getGroupSet.toArray ++ auxGroupSet
   }
 
   /**
